requirements.txt
yt-dlp>=2024.12.10
spotipy>=2.22.1
mutagen>=1.46.0
tqdm>=4.65.0
python-dotenv>=1.0.0
requests>=2.31.0
